add overflow checker + move logging code
